commit
commit

+++ src/main/java/tn/esprit/spring/repository/IActifFinancier.java
@@ -23,10 +23,10 @@ List<ActifFinancier> findByNomFond(Fond nom_fond) ;
 	public int listEuro_Croissance( );
 
 	@Query(value = "SELECT * FROM actif_financier WHERE nom_fond='Euro_Croissance' and  etat='En cours' and user_actif_user_id=:w ",nativeQuery = true)
-	 List<ActifFinancier> listemontant_actuelEuroCroissanceparUser( @Param("w") int w );
+	 List<ActifFinancier> listemontant_actuelEuroCroissanceparUser( @Param("w") Long w );
 
 	@Query(value = "SELECT * FROM actif_financier WHERE nom_fond='Fond_Euro' and  etat='En cours' and user_actif_user_id=:w ",nativeQuery = true)
-	 List<ActifFinancier> listemontant_actuelFondEuroparUser( @Param("w") int w );
+	 List<ActifFinancier> listemontant_actuelFondEuroparUser( @Param("w") Long w );
 	
 	
 	List<ActifFinancier> findByUserActif( User user);
